ECS physics system works as expected.

+++ core/src/com/mygdx/game/model/game/component/ScoreComponent.java
@@ -2,24 +2,24 @@ package com.mygdx.game.model.game.component;
 
 public class ScoreComponent extends Component
 {
-    private  int value;
+    private float value;
     public  ScoreComponent(int value)
     {
         this.value = value;
     }
 
-    public int getValue() {
+    public float getValue() {
         return this.value;
     }
 
-    public void decrease(int value)
+    public void decrease(float value)
     {
         if(value <= 0)
             throw new IllegalArgumentException("Invalid decreasement value");
         this.value = Math.max(0, this.value - value);
     }
 
-    public void increase(int value)
+    public void increase(float value)
     {
         if(value <= 0)
             throw new IllegalArgumentException("Invalid increasement value");
